refactor: replace Modals component with individual modal components
- Replace single Modals.vue with individual modal components (LogoutModal, ClearLogsModal, AssignStudentModal, StudentInfoModal)
- Implement dynamic modal state management in AdminView
- Remove unused layout store
- Update modal handling in Header component
