+++ src/views/AdminView.vue
@@ -1,12 +1,31 @@
 <script setup lang="ts">
 // IMPORTS
-import { defineAsyncComponent } from 'vue'
+import { ref, provide, defineAsyncComponent } from 'vue'
 import { RouterView } from 'vue-router'
 
 // LAYOUT COMPONENTS
 const LayoutHeader = defineAsyncComponent(() => import('@/components/layout/Header.vue'))
 const AdminSidebar = defineAsyncComponent(() => import('@/components/layout/Sidebar.vue'))
-const AdminModals = defineAsyncComponent(() => import('@/components/modals/Modals.vue'))
+
+// MODAL COMPONENTS
+const LogoutModal = defineAsyncComponent(() => import('@/components/modals/LogoutModal.vue'))
+const ClearLogsModal = defineAsyncComponent(() => import('@/components/modals/ClearLogsModal.vue'))
+const AssignStudentModal = defineAsyncComponent(() => import('@/components/modals/AssignStudentModal.vue'))
+const StudentInfoModal = defineAsyncComponent(() => import('@/components/modals/StudentInfoModal.vue'))
+
+// MODAL STATE
+const activeModal = ref<string | null>(null)
+
+const showModal = (modalName: string) => {
+  activeModal.value = modalName
+}
+
+const hideModal = () => {
+  activeModal.value = null
+}
+
+// PROVIDE MODAL METHODS
+provide('showModal', showModal)
 </script>
 
 <template>
@@ -29,7 +48,10 @@ const AdminModals = defineAsyncComponent(() => import('@/components/modals/Modal
     </div>
 
     <!-- MODALS -->
-    <AdminModals />
+    <LogoutModal v-if="activeModal === 'logout'" @close="hideModal" />
+    <ClearLogsModal v-if="activeModal === 'clear-logs'" @close="hideModal" />
+    <AssignStudentModal v-if="activeModal === 'assign-student'" @close="hideModal" />
+    <StudentInfoModal v-if="activeModal === 'student-info'" @close="hideModal" />
   </div>
 </template>
